Build entire log line at once
Not quite entirely; we still generate the attribute list in a separate
buffer, but once we have all the pieces, we glue them all together with
one final Sprintf statement at the end. This is to support overriding
some fields (i.e., the caller) in the attribute list, so we have to wait
until after we've processed the attribute list before we generate the
final log line.