Include indirect package files in linter
The "revive ./..." command only analyzes the files that are directly
part of the package. It doesn't analyze the Mage files or the tests.
This commit determines the remaining files and includes them in the
revive command line. It also addresses all the issues reported in the
auxiliary source files, either by simplifying the code or by disabling
some of the rules that aren't so useful in tests.