Ignore argument limit for attrToJSON
A bunch of the arguments to this function are for managing the state of
the JSON serializer. The "right" solution would be to provide a class
that keeps track of the current state and "knows" what additional output
it needs to generate before serializing the next data. But that's a
bunch of work I don't want to do right now, so instead we'll just ignore
the linter when it complains about all the arguments.