Switch back to option-function style
Having a single struct with all the options was nice for showing what
options were available, but because of limitations in the documentation
formatter, it was hard to provide good-looking examples for each option.
(Examples associated with _fields_ are not displayed, neither by pkgsite
nor by godocs.)

Using an option function for each possible option allows formatted
documentation for each option (instead of just displaying the comments
verbatim in a code block) and for examples tailored to each option.