Rewrite handler with different recursion strategy
This introduces three different "node" types for a chain of legacy log
handlers. The base, group, and addtribute handlers each have specific
behavior for certain portions of the logging process, and having
separate types for each makes it easier to know which task to do. For
example, an attribute node _only_ stores a non-empty list of attributes,
so there's never a question of whether attributes need to be logged.