Move readme to separate directory
Pkgsite includes an excerpt of the project readme at the top of the
generated documentation, where it ends up being largely redundant with
the package documentation. But GitHub will display a readme from a
"docs" subdirectory, so we can essentially "hide" the readme from
pkgsite while still displaying it on the GitHub page. This also avoids
having a redundant pkgsite link displayed on the pkgsite page.